api/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]==2.5.3

# YAML processing
PyYAML==6.0.1

# Additional utilities
python-multipart==0.0.6  # For form data
python-jose[cryptography]==3.3.0  # For JWT tokens (future)
passlib[bcrypt]==1.7.4  # For password hashing (future)

# Rate limiting
slowapi==0.1.9

# Environment variables
python-dotenv==1.0.0

# CORS
# (included in FastAPI)

# Development tools (optional)
httpx==0.26.0  # For testing
pytest==7.4.4
pytest-asyncio==0.23.3